requirements.txt
confluent-kafka
requests==2.31.0
psycopg2-binary==2.9.10
cassandra-driver==3.29.0
python-dotenv==1.0.0
